--- client/app/views/alert.vid.tmpl.html
@@ -1,21 +1,0 @@
-<br>
-<input ng-model="alertSearch" class="form-control bg-primary" placeholder="Search Alert details..."></lebel>
-<table class="table table-bordered ">
-    <thead>
-    <tr>
-        <th>Vin</th>
-        <th>TimeStamp</th>
-        <th>Priority</th>
-        <th>Message</th>
-        <th>Location </th>
-    </tr>
-    </thead>
-    <tr ng-repeat="alert in alertVidVm.alerts | filter:alertSearch"ng-class="{'HIGH': 'bg-danger' ,'LOW':'bg-success','MEDIUM':'bg-warning'}[alert.priority]">
-        <td>{{alert.vin}}</a></td>
-        <td>{{alert.timestamp |date:"MM/dd/yyyy 'at' h:mma"}}</td>
-        <td>{{alert.priority}}</td>
-        <td>{{alert.alertmsg}}</td>
-        <td><a ng-href="#!/alerts/map/{{alert.alertid}}">location</a></td>
-    </tr>
-</table>
-
