+++ client/bower.json
@@ -16,13 +16,13 @@
     "tests"
   ],
   "dependencies": {
-    "angular": "^1.6.4",
-    "bootstrap": "^3.3.7",
-    "jquery": "^3.2.1",
-    "angular-route": "^1.6.4",
-    "angular-messages": "^1.6.4",
-    "angular-bootstrap": "^2.5.0",
-    "ngmap": "^1.18.4"
+    "angular": "~1.6.4",
+    "bootstrap": "~3.3.7",
+    "jquery": "~3.2.1",
+    "angular-route": "~1.6.4",
+    "angular-messages": "~1.6.4",
+    "angular-bootstrap": "~2.5.0",
+    "ngmap": "~1.18.4"
   },
   "overrides": {
     "bootstrap": {
@@ -38,5 +38,8 @@
     "jquery": {
       "main": []
     }
+  },
+  "resolutions": {
+    "angular": "~1.6.4"
   }
 }
