+++ client/package.json
@@ -1,19 +1,24 @@
 {
-  "name": "ui-test",
+  "name": "car-tracker",
   "version": "0.0.1",
-  "description": "\"test UI\"",
+  "description": "\"Auto Tracker which consumes rest end point powered by angular\"",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "shesh",
+  "author": "sheshadri",
   "license": "ISC",
   "devDependencies": {
     "browser-sync": "^2.18.12",
     "gulp": "^3.9.1",
     "gulp-angular-filesort": "^1.1.1",
     "gulp-clean": "^0.3.2",
+    "gulp-clean-css": "^3.4.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-filter": "^5.0.0",
     "gulp-inject": "^4.2.0",
-    "main-bower-files": "^2.13.1"
+    "gulp-uglify": "^3.0.0",
+    "main-bower-files": "^2.13.1",
+    "merge-stream": "^1.0.1"
   }
 }
