+++ client/src/app/app.constant.js
@@ -4,7 +4,9 @@
             'API_HOST' :'http://localhost:8080/tracker/api/',
             'API_VEHICLE' :'http://localhost:8080/tracker/api/vehicles',
             'API_ALERT' :'http://localhost:8080/tracker/api/alerts/',
-            'API_READINGS' :'http://localhost:8080/tracker/api/readings',
-            'API_VEHICLE_ID' :'http://localhost:8080/tracker/api/vehicle/'
+            'API_READINGS' :'http://localhost:8080/tracker/api/readings/',
+            'API_VEHICLE_ID' :'http://localhost:8080/tracker/api/vehicle/',
+            'IMAGE_HOME' : './src/images',
+            'GOOGLE_KEY' : 'AIzaSyAEg5-KmpXA06XscGZyZs6wdh-QvIuJbWM'
         });
 })();
