+++ client/src/app/controller/vehicle.controller.js
@@ -3,10 +3,14 @@
     angular.module('app')
         .controller('vehicleController',vehicleControllerFn);
 
-    vehicleControllerFn.$inject =['vehicleService'];
-    function vehicleControllerFn(vehicleService) {
+    vehicleControllerFn.$inject =['vehicleService','$rootScope'];
+
+    function vehicleControllerFn(vehicleService,$rootScope) {
         var vehicleVm  = this;
+         $rootScope.flag = false;
+
         console.log("Vehicle Controller Invoked");
+        console.log($rootScope.flag);
         vehicleService.getAll()
             .then(function (vehicles) {
                     console.log("In controller The data is found")
