+++ client/src/app/services/map.service.js
