+++ client/src/app/views/alerts/alert.all.tmpl.html
