+++ client/src/app/views/alerts/alert.map.tmpl.html
