+++ client/src/app/views/readings/readings.list.tmpl.html
