+++ client/src/app/views/readings/readings.vin.tmpl.html
