+++ client/src/app/views/vehicles/vehicle.attributes.tmpl.html
@@ -7,7 +7,7 @@
  </ng-map>
 </div>
 <br>
-<a class="btn btn-success" href ="#!/vehicle/reading/{{vehicleAttVm.vin}}" ng-click="vehicleAttVm.toggleReading()"> Readings  </a>
+<a class="btn btn-info text-danger" href ="#!/vehicle/reading/{{vehicleAttVm.vin}}" ng-click="vehicleAttVm.toggleReading()"> Readings  </a>
 <br>
 <div ng-show="vehicleAttVm.readingFlag" >
     <br>
