+++ client/src/app/views/vehicles/vehicle.detail.tmpl.html
