+++ client/src/app/views/vehicles/vehicle.list.tmpl.html
