+++ client/src/index.html
@@ -1,10 +1,16 @@
 <!DOCTYPE html>
-<html lang="en" >
+<html lang="en" ng-app="app"  >
 <head>
 
     <meta charset="UTF-8">
     <!--bower:css -->
-    <link rel="stylesheet" href="../bower_components/bootstrap/dist/css/bootstrap.css">
+    <link rel="stylesheet" href="/bower_components/bootstrap/dist/css/bootstrap.css">
+    <!--endinject-->
+
+    <!--vendor:css -->
+    <!--endinject-->
+
+    <!--app:css -->
     <!--endinject-->
 
     <!--inject:css -->
@@ -16,76 +22,84 @@
     <link rel="stylesheet" href="styles/readings.css">
     <link rel="stylesheet" href="styles/vehicle.css">
     <!--endinject-->
-    <title>Demo</title>
-</head>
-<body ng-app="app" class="container" >
 
-<header class="container">
-    <table><tr>
-        <td><img src="images/ferrari1.jpeg" class="img-rounded" width="110px" height="80px"></td>
-        <td><img src="images/ferrari2.png" class="img-rounded " width="120px" height="80px"></td>
-        <td><img src="images/merc.png" class="img-rounded " width="120px" height="80px"></td>
-        <td><img src="images/rr.jpeg" class="img-rounded " width="120px" height="80px"></td>
-        <td><img src="images/lambo.jpeg" class="img-rounded " width="110px" height="80px"></td>
+
+    <title>Car Tracker project</title>
+</head>
+
+<body class="container" >
+<header>
+    <div class="container">
+    <table ><tr>
+        <td align="left"><img src="images/angred.png" class="img-rounded " width="200px" height="100px"></td>
+        <td class="text-center text-danger  glyphicon"> <h3>AUTO TRACKER</h3></td>
     </tr></table>
+        <ul class="nav nav-pills  navbar-inverse">
+            <li class="nav-item bg-danger">
+                <a class="nav-link active" href="#!home">Home</a>
+            </li>
+            <li class="nav-item bg-danger">
+                <a class="nav-link" href="#!vehicles">Vehicle</a>
+            </li>
+            <li class="nav-item bg-danger">
+                <a class="nav-link" href="#!alerts">Alerts</a>
+            </li>
+            <li class="nav-item bg-danger">
+                <a class="nav-link" href="#!readings">Readings</a>
+            </li>
+        </ul>
+    </div>
 </header>
 <br>
 
-<ul class="nav nav-pills  navbar-inverse">
-    <li class="nav-item bg-danger">
-        <a class="nav-link" href="#!home">Home</a>
-    </li>
-    <li class="nav-item bg-danger">
-        <a class="nav-link active" href="#!vehicles">Vehicle</a>
-    </li>
-    <li class="nav-item bg-danger">
-        <a class="nav-link" href="#!alerts">Alerts</a>
-    </li>
-</ul>
-
-
-
 <!--bower:js -->
-<script src="../bower_components/angular/angular.js"></script>
-<script src="../bower_components/angular-route/angular-route.js"></script>
-<script src="../bower_components/angular-messages/angular-messages.js"></script>
-<script src="../bower_components/angular-bootstrap/ui-bootstrap-tpls.js"></script>
-<script src="../bower_components/ngmap/build/scripts/ng-map.js"></script>
-<script src="https://maps.googleapis.com/maps/api/js?libraries=places&key=AIzaSyAEg5-KmpXA06XscGZyZs6wdh-QvIuJbWM"></script>
+<script src="/bower_components/angular/angular.js"></script>
+<script src="/bower_components/angular-route/angular-route.js"></script>
+<script src="/bower_components/angular-messages/angular-messages.js"></script>
+<script src="/bower_components/angular-bootstrap/ui-bootstrap-tpls.js"></script>
+<script src="/bower_components/ngmap/build/scripts/ng-map.js"></script>
 <!--endinject-->
+
+
 
 <!--inject:js -->
-
 <script src="app/app.js"></script>
-
-<script src="app/services/vehicle.service.js"></script>
-
-<script src="app/services/reading.service.js"></script>
-
-<script src="app/services/alert.service.js"></script>
-
 <script src="app/controller/vehicle.detail.controller.js"></script>
-
 <script src="app/controller/vehicle.controller.js"></script>
-
+<script src="app/controller/vehicle.attributes.controller.js"></script>
+<script src="app/controller/readings.controller.vin.js"></script>
 <script src="app/controller/readings.controller.js"></script>
-
+<script src="app/controller/home.controller.js"></script>
 <script src="app/controller/alert.map.controller.js"></script>
-
+<script src="app/controller/alert.controller.vin.js"></script>
 <script src="app/controller/alert.controller.js"></script>
-
-<script src="app/controller/alert.controller.byVid.js"></script>
-
+<script src="app/services/vehicle.service.js"></script>
+<script src="app/services/reading.service.js"></script>
+<script src="app/services/map.service.js"></script>
+<script src="app/services/alert.service.js"></script>
 <script src="app/app.constant.js"></script>
-
 <!--endinject-->
-<section id="section">
-    <div ng-view> </div>
+
+
+<!--vendor:js -->
+<!--endinject-->
+
+<!--app:js -->
+<!--endinject-->
+<script src="https://maps.googleapis.com/maps/api/js?libraries=places&key=AIzaSyAEg5-KmpXA06XscGZyZs6wdh-QvIuJbWM"></script>
+
+<section class="container" >
+    <div ng-view >
+    </div>
+    <img src="images/car1.jpeg" width="100%" height="100%" >
+
+
 </section>
-<footer id="footer">
-    <div>
-        <p  align="centre"> &copy; Auto Tracker - Demo Project </p>
+
+</body>
+<footer class="container" >
+    <div class="navbar-inverse">
+        <h3 align="center" class="text-danger"> &copy; Auto Tracker - Demo Project </h3>
     </div>
 </footer>
-</body>
 </html>
