--- client/styles/alert.css
@@ -1,4 +1,0 @@
-#map{
-    height: 500px;
-    width: 500px;
-}
