--- client/styles/common.css
@@ -1,16 +1,0 @@
-.section{
-    display: none;
-    min-height: 70%;
-    max-height: 80%;
-    background: lightgrey;
-}
-#slides_control > div{
-    height: 200px;
-}
-#slides_control {
-    position:absolute;
-    width: 400px;
-    left:50%;
-    top:20px;
-    margin-left:-200px;
-}
