--- client/styles/footer.css
@@ -1,5 +1,0 @@
-#footer{
-    min-height: 40px;
-    border-bottom: hidden;
-    background: white;
-}
