--- client/styles/header.css
@@ -1,4 +1,0 @@
-#header{
-    min-height: 40px;
-    background:white;
-}
